backend/requirements.txt
alembic==1.8.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
